[NO ISSUE]: Fix failing CloudStorageGCSTest
- user model changes: no
- storage format changes: no
- interface changes: no

details:
- cc-cloud-storage-gcs.conf has been misconfigured to use
  5 bytes instead of 5MB
- Avoid unncessary cast to int

Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/19346
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>

+++ asterixdb/asterix-cloud/src/main/java/org/apache/asterix/cloud/clients/google/gcs/GCSWriter.java
@@ -66,8 +66,8 @@ public class GCSWriter implements ICloudWriter {
         guardian.checkIsolatedWriteAccess(bucket, path);
         // The GCS library triggers a new upload when its internal buffer is full, not on each call to writer.write().
         // uploadsToBeTriggered estimates upload count, and we acquire matching tokens from the limiter.
-        int uploadsToBeTriggered =
-                (int) ((writtenBytes + page.remaining()) / writeBufferSize) - (int) (writtenBytes / writeBufferSize);
+        long uploadsToBeTriggered =
+                ((writtenBytes + page.remaining()) / writeBufferSize) - (writtenBytes / writeBufferSize);
         while (uploadsToBeTriggered-- > 0) {
             profiler.objectMultipartUpload();
         }
